new way to run clusters with numNodes and type of cluster as params

+++ run.sh
@@ -1,22 +1,25 @@
 #!/bin/bash
 
-if [ $# -ne 1 ] ; then
-  echo "Usage: ./run.sh [centos|ubuntu]"
-  echo ""
-  exit 1
-fi
-
-if [ "$1" == "centos" ] ; then
- cp Vagrantfile.centos Vagrantfile
-elif [ "$1" == "ubuntu" ] ; then
- cp Vagrantfile.ubuntu Vagrantfile
-else
-  echo "Usage: ./run.sh [centos|ubuntu]"
+if [ $# -ne 3 ] ; then
+  echo "Usage: ./run.sh [centos|ubuntu] [1|3] [ndb|hopsworks|hopsworks-ha]"
   echo ""
   exit 1
 fi
 
 set -e
+
+if [ ! -f Vagrantfile.$1.$2 ] ; then
+ echo "Couldn't find the Vagrantfile.$1.$2 for your cluster"
+ exit 1
+fi
+if [ ! -f cluster.yml.$2.$3 ] ; then
+ echo "Couldn't find the Vagrantfile.$1.$2 for your cluster"
+ exit 1
+fi
+ 
+cp Vagrantfile.$1.$2 Vagrantfile
+cp cluster.yml.$2.$3 cluster.yml
+
 echo "Removing old vendored cookbooks"
 rm -rf cookbooks > /dev/null 2>&1
 rm -f Berksfile.lock nohup.out > /dev/null 2>&1
